Refactor area member and owner management
Unified member and owner handling by moving related methods to the CraftArea superclass. This reduces code duplication and centralizes the logic for consistency and easier maintenance.
